fix: endpoints
